+++ src/test/java/ch/ethz/matsim/discrete_mode_choice/components/tour_finder/ActivityTourFinderTest.java
@@ -34,7 +34,7 @@ public class ActivityTourFinderTest {
 		for (Trip trip : trips) {
 			String initialMode = trip.getLegsOnly().get(0).getMode();
 			modeChoiceTrips.add(new DiscreteModeChoiceTrip(trip.getOriginActivity(), trip.getDestinationActivity(),
-					initialMode, 0.0, 0, 0));
+					initialMode, trip.getTripElements(), 0.0, 0, 0));
 		}
 
 		return modeChoiceTrips;